rename example to ExampleApp
