+++ ExampleApp/AppShell.xaml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <Shell
-    x:Class="example.AppShell"
+    x:Class="ExampleApp.AppShell"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:example"
+    xmlns:local="clr-namespace:ExampleApp"
     Shell.FlyoutBehavior="Disabled"
-    Title="example">
+    Title="ExampleApp">
 
     <ShellContent
         Title="Home"
