+++ ExampleApp/Platforms/Android/AndroidManifest.xml
