+++ ExampleApp/Platforms/Android/MainApplication.cs
@@ -1,7 +1,7 @@
 ﻿using Android.App;
 using Android.Runtime;
 
-namespace example;
+namespace ExampleApp;
 
 [Application]
 public class MainApplication : MauiApplication
