+++ ExampleApp/Platforms/Android/Resources/values/colors.xml
