+++ ExampleApp/Platforms/MacCatalyst/Entitlements.plist
