+++ ExampleApp/Platforms/MacCatalyst/Info.plist
