+++ ExampleApp/Platforms/Tizen/Main.cs
@@ -2,7 +2,7 @@ using System;
 using Microsoft.Maui;
 using Microsoft.Maui.Hosting;
 
-namespace example;
+namespace ExampleApp;
 
 class Program : MauiApplication
 {
