+++ ExampleApp/Platforms/Windows/App.xaml
@@ -1,8 +1,8 @@
 ﻿<maui:MauiWinUIApplication
-    x:Class="example.WinUI.App"
+    x:Class="ExampleApp.WinUI.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:maui="using:Microsoft.Maui"
-    xmlns:local="using:example.WinUI">
+    xmlns:local="using:ExampleApp.WinUI">
 
 </maui:MauiWinUIApplication>
