+++ ExampleApp/Platforms/Windows/Package.appxmanifest
