+++ ExampleApp/Platforms/iOS/Resources/PrivacyInfo.xcprivacy
