+++ ExampleApp/Resources/Images/dotnet_bot.png
