+++ ExampleApp/Resources/Raw/AboutAssets.txt
