+++ ExampleApp/Resources/Styles/Colors.xaml
