+++ ExampleApp/Resources/Styles/Styles.xaml
